+++ changelog.md
@@ -1,6 +1,6 @@
 # next tasks
-1. latest invoices in dashboard
-2. latest clients in dashboard
+1. latest clients in dashboard
+2. invalidate/modify pages when invoice is created.
 3. clients should be also sortable and paginable
 4. Refactor/organize based on [slide for Week 5 Office Hour 1](https://docs.google.com/presentation/d/1pLZ8Zvpyr_myotpvyVYzY5awpUFUO78DE5CnBnVNm7c/edit#slide=id.g117828bad90_0_9)
 5. latest client invoices in profile (link menu item from client table)
@@ -18,6 +18,7 @@
   - throwing exception here: invoiceslice: thunkAPI.dispatch(beforeUpdate(prevInvoice));
   - cold loading client page, e.g. http://localhost:3000/client-dashboard?id=18e50f68-0b4c-4c2a-9bdf-5fe9dcedf3ae
   - NOT WORKIGN.... double check :when updating invoice details, the corresponging filtered pages should also be updated. 
+- Added latest invoices to dashboard
 
 # 2022 03 20
 - cancel buttons should go back to previous page